# Conflicts:
#	.env
#	app/Http/Controllers/InvoiceController.php
#	package-lock.json
#	public/css/app.css
#	public/js/app.js
#	public/mix-manifest.json
#	routes/web.php

+++ database/migrations/2009_09_06_090235_create_companies_table.php
@@ -17,11 +17,12 @@ class CreateCompaniesTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('address');
-            $table->float('wallet')->default(0.00);
+            $table->double('wallet')->default(0.00);
             $table->string('postal_code');
             $table->string('email');
             $table->string('phone_number');
             $table->timestamps();
+            
         });
     }
